rename models -> enums
