+++ app/enums/fuckult.py
