+++ app/enums/week.py
